fixed build bug
the liquid exception \\today invalid time is because in the handouts there is a yml metadata date which is \today for pandoc latex to use. But Jekyll thinks all markdown files are pages so there is the liquid exception. Solution is to add underscore to handout folder to let jekyll ignore the folder when building site.